[Security Solution][Alert details] - move table tab content to flyout folder
