--- x-pack/plugins/security_solution/public/common/components/event_details/table/field_value_cell.tsx
@@ -1,93 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-import type { CSSProperties } from 'react';
-import React from 'react';
-import { EuiFlexGroup, EuiFlexItem, EuiText } from '@elastic/eui';
-import type { BrowserField } from '../../../containers/source';
-import { OverflowField } from '../../tables/helpers';
-import { FormattedFieldValue } from '../../../../timelines/components/timeline/body/renderers/formatted_field';
-import { MESSAGE_FIELD_NAME } from '../../../../timelines/components/timeline/body/renderers/constants';
-import type { EventFieldsData, FieldsData } from '../types';
-import { getFieldFormat } from '../get_field_format';
-
-export interface FieldValueCellProps {
-  contextId: string;
-  data: EventFieldsData | FieldsData;
-  eventId: string;
-  fieldFromBrowserField?: Partial<BrowserField>;
-  getLinkValue?: (field: string) => string | null;
-  isDraggable?: boolean;
-  linkValue?: string | null | undefined;
-  style?: CSSProperties | undefined;
-  values: string[] | null | undefined;
-}
-
-export const FieldValueCell = React.memo(
-  ({
-    contextId,
-    data,
-    eventId,
-    fieldFromBrowserField,
-    getLinkValue,
-    isDraggable = false,
-    linkValue,
-    style,
-    values,
-  }: FieldValueCellProps) => {
-    return (
-      <EuiFlexGroup
-        alignItems="flexStart"
-        data-test-subj={`event-field-${data.field}`}
-        direction="column"
-        gutterSize="none"
-        style={style}
-      >
-        {values != null &&
-          values.map((value, i) => {
-            if (fieldFromBrowserField == null) {
-              return (
-                <EuiFlexItem grow={false} key={`${i}-${value}`}>
-                  <EuiText size="xs" key={`${i}-${value}`}>
-                    {value}
-                  </EuiText>
-                </EuiFlexItem>
-              );
-            }
-            return (
-              <EuiFlexItem
-                className="eventFieldsTable__fieldValue"
-                grow={false}
-                key={`${i}-${value}`}
-              >
-                {data.field === MESSAGE_FIELD_NAME ? (
-                  <OverflowField value={value} />
-                ) : (
-                  <FormattedFieldValue
-                    contextId={`${contextId}-${eventId}-${data.field}-${i}-${value}`}
-                    eventId={eventId}
-                    fieldFormat={getFieldFormat(data)}
-                    fieldName={data.field}
-                    fieldFromBrowserField={fieldFromBrowserField}
-                    fieldType={data.type}
-                    isAggregatable={fieldFromBrowserField.aggregatable}
-                    isDraggable={isDraggable}
-                    isObjectArray={data.isObjectArray}
-                    value={value}
-                    linkValue={(getLinkValue && getLinkValue(data.field)) ?? linkValue}
-                    truncate={false}
-                  />
-                )}
-              </EuiFlexItem>
-            );
-          })}
-      </EuiFlexGroup>
-    );
-  }
-);
-
-FieldValueCell.displayName = 'FieldValueCell';
