+++ x-pack/plugins/security_solution/public/flyout/document_details/right/tabs/table_tab.tsx
@@ -5,46 +5,115 @@
  * 2.0.
  */
 
-import React, { memo } from 'react';
-import { EuiText } from '@elastic/eui';
-import { getFieldFromBrowserField } from '../../../../common/components/event_details/columns';
+import React, { memo, useCallback, useMemo, useState } from 'react';
+import { getOr, sortBy } from 'lodash/fp';
+import memoizeOne from 'memoize-one';
+import styled from 'styled-components';
+import { type EuiBasicTableColumn, EuiText, EuiInMemoryTable } from '@elastic/eui';
+import { i18n } from '@kbn/i18n';
+import { dataTableSelectors, tableDefaults } from '@kbn/securitysolution-data-table';
+import { getCategory } from '@kbn/triggers-actions-ui-plugin/public';
+import type {
+  BrowserField,
+  BrowserFields,
+  TimelineEventsDetailsItem,
+} from '@kbn/timelines-plugin/common';
+import { TableFieldNameCell } from '../components/table_field_name_cell';
+import { TableFieldValueCell } from '../components/table_field_value_cell';
+import { TABLE_TAB_CONTENT_TEST_ID, TABLE_TAB_SEARCH_INPUT_TEST_ID } from './test_ids';
+import { getAllFieldsByName } from '../../../../common/containers/source';
+import { useDeepEqualSelector } from '../../../../common/hooks/use_selector';
+import { timelineDefaults } from '../../../../timelines/store/defaults';
+import { timelineSelectors } from '../../../../timelines/store';
 import type { EventFieldsData } from '../../../../common/components/event_details/types';
-import { FieldValueCell } from '../../../../common/components/event_details/table/field_value_cell';
-import { FieldNameCell } from '../../../../common/components/event_details/table/field_name_cell';
 import { CellActions } from '../components/cell_actions';
-import * as i18n from '../../../../common/components/event_details/translations';
 import { useDocumentDetailsContext } from '../../shared/context';
-import type { ColumnsProvider } from '../../../../common/components/event_details/event_fields_browser';
-import { EventFieldsBrowser } from '../../../../common/components/event_details/event_fields_browser';
-import { TimelineTabs } from '../../../../../common/types';
+import { isInTableScope, isTimelineScope } from '../../../../helpers';
 
-export const getColumns: ColumnsProvider = ({
-  browserFields,
-  eventId,
-  contextId,
-  scopeId,
-  getLinkValue,
-  isDraggable,
-}) => [
+/**
+ * As the flyout is quite narrow, we reduce the font size to fit more content
+ */
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+const StyledEuiMemoryTable = styled(EuiInMemoryTable as any)`
+  .flyout-table-row-small-font {
+    font-size: ${({ theme }) => theme.eui.euiFontSizeXS};
+    font-family: ${({ theme }) => theme.eui.euiCodeFontFamily};
+  }
+`;
+
+const COUNT_PER_PAGE_OPTIONS = [25, 50, 100];
+
+const PLACEHOLDER = i18n.translate('xpack.securitySolution.flyout.table.filterPlaceholderLabel', {
+  defaultMessage: 'Filter by field or value...',
+});
+export const FIELD = i18n.translate('xpack.securitySolution.flyout.table.fieldCellLabel', {
+  defaultMessage: 'Field',
+});
+const VALUE = i18n.translate('xpack.securitySolution.flyout.table.valueCellLabel', {
+  defaultMessage: 'Value',
+});
+
+/**
+ * Defines the behavior of the search input that appears above the table of data
+ */
+const search = {
+  box: {
+    incremental: true,
+    placeholder: PLACEHOLDER,
+    schema: true,
+    'data-test-subj': TABLE_TAB_SEARCH_INPUT_TEST_ID,
+  },
+};
+
+/**
+ * Retrieve the correct field from the BrowserField
+ */
+export const getFieldFromBrowserField = memoizeOne(
+  (field: string, browserFields: BrowserFields): BrowserField | undefined => {
+    const category = getCategory(field);
+
+    return browserFields[category]?.fields?.[field] as BrowserField;
+  },
+  (newArgs, lastArgs) => newArgs[0] === lastArgs[0]
+);
+
+export type ColumnsProvider = (providerOptions: {
+  /**
+   * An object containing fields by type
+   */
+  browserFields: BrowserFields;
+  /**
+   * Id of the document
+   */
+  eventId: string;
+  /**
+   * Maintain backwards compatibility // TODO remove when possible
+   */
+  scopeId: string;
+  /**
+   * Value of the link field if it exists. Allows to navigate to other pages like host, user, network...
+   */
+  getLinkValue: (field: string) => string | null;
+}) => Array<EuiBasicTableColumn<TimelineEventsDetailsItem>>;
+
+export const getColumns: ColumnsProvider = ({ browserFields, eventId, scopeId, getLinkValue }) => [
   {
     field: 'field',
     name: (
       <EuiText size="xs">
-        <strong>{i18n.FIELD}</strong>
+        <strong>{FIELD}</strong>
       </EuiText>
     ),
     width: '30%',
     render: (field, data) => {
-      return (
-        <FieldNameCell data={data as EventFieldsData} field={field} fieldMapping={undefined} />
-      );
+      return <TableFieldNameCell dataType={(data as EventFieldsData).type} field={field} />;
     },
   },
   {
     field: 'values',
     name: (
       <EuiText size="xs">
-        <strong>{i18n.VALUE}</strong>
+        <strong>{VALUE}</strong>
       </EuiText>
     ),
     width: '70%',
@@ -52,13 +121,12 @@ export const getColumns: ColumnsProvider = ({
       const fieldFromBrowserField = getFieldFromBrowserField(data.field, browserFields);
       return (
         <CellActions field={data.field} value={values} isObjectArray={data.isObjectArray}>
-          <FieldValueCell
-            contextId={contextId}
+          <TableFieldValueCell
+            contextId={scopeId}
             data={data as EventFieldsData}
             eventId={eventId}
             fieldFromBrowserField={fieldFromBrowserField}
             getLinkValue={getLinkValue}
-            isDraggable={isDraggable}
             values={values}
           />
         </CellActions>
@@ -68,23 +136,99 @@ export const getColumns: ColumnsProvider = ({
 ];
 
 /**
- * Table view displayed in the document details expandable flyout right section
+ * Table view displayed in the document details expandable flyout right section Table tab
  */
-// TODO: MOVE TO FLYOUT FOLDER - https://github.com/elastic/security-team/issues/7462
 export const TableTab = memo(() => {
   const { browserFields, dataFormattedForFieldBrowser, eventId, scopeId } =
     useDocumentDetailsContext();
 
+  const [pagination, setPagination] = useState<{ pageIndex: number }>({
+    pageIndex: 0,
+  });
+  const onTableChange = useCallback(({ page: { index } }: { page: { index: number } }) => {
+    setPagination({ pageIndex: index });
+  }, []);
+
+  const getScope = useMemo(() => {
+    if (isTimelineScope(scopeId)) {
+      return timelineSelectors.getTimelineByIdSelector();
+    } else if (isInTableScope(scopeId)) {
+      return dataTableSelectors.getTableByIdSelector();
+    }
+  }, [scopeId]);
+
+  const defaults = useMemo(
+    () => (isTimelineScope(scopeId) ? timelineDefaults : tableDefaults),
+    [scopeId]
+  );
+
+  const columnHeaders = useDeepEqualSelector((state) => {
+    const { columns } = (getScope && getScope(state, scopeId)) ?? defaults;
+    return columns;
+  });
+
+  const fieldsByName = useMemo(() => getAllFieldsByName(browserFields), [browserFields]);
+
+  const items = useMemo(
+    () =>
+      sortBy(['field'], dataFormattedForFieldBrowser).map((item, i) => ({
+        ...item,
+        ...fieldsByName[item.field],
+        valuesConcatenated: item.values != null ? item.values.join() : '',
+        ariaRowindex: i + 1,
+      })),
+    [dataFormattedForFieldBrowser, fieldsByName]
+  );
+
+  const getLinkValue = useCallback(
+    (field: string) => {
+      const columnHeader = columnHeaders.find((col) => col.id === field);
+      if (!columnHeader || !columnHeader.linkField) {
+        return null;
+      }
+      const linkFieldData = (dataFormattedForFieldBrowser ?? []).find(
+        (d) => d.field === columnHeader.linkField
+      );
+      const linkFieldValue = getOr(null, 'originalValue', linkFieldData);
+      return Array.isArray(linkFieldValue) ? linkFieldValue[0] : linkFieldValue;
+    },
+    [dataFormattedForFieldBrowser, columnHeaders]
+  );
+
+  // forces the rows of the table to render smaller fonts
+  const onSetRowProps = useCallback(
+    ({ field }: TimelineEventsDetailsItem) => ({
+      className: 'flyout-table-row-small-font',
+      'data-test-subj': `flyout-table-row-${field}`,
+    }),
+    []
+  );
+
+  const columns = useMemo(
+    () =>
+      getColumns({
+        browserFields,
+        eventId,
+        scopeId,
+        getLinkValue,
+      }),
+    [browserFields, eventId, scopeId, getLinkValue]
+  );
+
   return (
-    <EventFieldsBrowser
-      browserFields={browserFields}
-      data={dataFormattedForFieldBrowser}
-      eventId={eventId}
-      isDraggable={false}
-      timelineTabType={TimelineTabs.query}
-      scopeId={scopeId}
-      isReadOnly={false}
-      columnsProvider={getColumns}
+    <StyledEuiMemoryTable
+      items={items}
+      itemId="field"
+      columns={columns}
+      onTableChange={onTableChange}
+      pagination={{
+        ...pagination,
+        pageSizeOptions: COUNT_PER_PAGE_OPTIONS,
+      }}
+      rowProps={onSetRowProps}
+      search={search}
+      sorting={false}
+      data-test-subj={TABLE_TAB_CONTENT_TEST_ID}
     />
   );
 });
