+++ x-pack/plugins/security_solution/public/flyout/document_details/right/tabs/test_ids.ts
@@ -7,6 +7,7 @@
 
 import { PREFIX } from '../../../shared/test_ids';
 
-export const TABLE_TAB_CONTENT_TEST_ID = 'event-fields-browser' as const;
+export const TABLE_TAB_CONTENT_TEST_ID = `${PREFIX}DocumentTable` as const;
+export const TABLE_TAB_SEARCH_INPUT_TEST_ID = `${PREFIX}DocumentTableSearchInput` as const;
 export const JSON_TAB_CONTENT_TEST_ID = 'jsonView' as const;
 export const JSON_TAB_COPY_TO_CLIPBOARD_BUTTON_TEST_ID = `${PREFIX}JsonTabCopyToClipboard` as const;
